requirements.txt
temporalio
terraform-local
gitpython
cdktf